[ft12] FIX unable to program phys. address of devices with address 0.0.0
Older Selfbus devices have the default KNX-address 0.0.0 after a clean firmware flash, which is not allowed and knxd will ignore the response, failing in programming a new phys. address.
This fix is not according to KNX spec.
In case of a APCI_INDIVIDUAL_ADDRESS_RESPONSE_PDU with sender and receiver address 0.0.0, we inject the sender address 15.15.255.
Possibly also a solution for the problem mentioned in Selfbus forum (https://selfbus.org/forum/viewtopic.php?f=2&t=429&start=10)