[RM] Fix behavior when both alarm and test alarm are requested via bus
Having both an alarm and a test alarm requested via bus is a strange
situation and should not happen, but it's important to behave sanely
even in such moments.

Previously, depending on the smoke detector behavior, we would either
continuously request the smoke detector to issue a test alarm, or
alternate between alarm and test alarm.

Now, when an alarm is requested we forward the alarm, and only the
alarm. Only when no alarm is requested, we consider forwarding the
test alarm.