[RM] Fix delayed alarm to continue on intermediate device status update
Previously, when another `deviceStatusUpdate` came in while the delayed
alarm counter was already running, the counter was restarted. That does
not make much sense; The counter should start when the alarm is reported
the first time and then continue counting down independent of other
device status updates that might come in later.