[RM] Fix periodic sending interval setting to be respected everywhere
Whenever a group object with periodic sending behavior is sent onto the
bus, restart the respective counter. This ensures reliable periodic
sending behavior.

This fix requires introduction of another counter for AlarmNetwork as
it can be sent independently of AlarmStatus when a delayed alarm is
configured.

Additionally, initialize these counters with their configured value so
we do not send bogus data onto the bus right after startup.