[RM] Implement burst prevention with random (0;2s) timer start delay
To prevent big bursts of info group object frames where some frames
might even need to be dropped, delay start the timer that triggers
device communication and info group object sending. The delay depends on
the device's individual address and is up to 2008 milliseconds long.
This ensures consecutive group objects do not overlap, and is sufficient
to allow other devices to send their frames in the "empty slots".

Closes #44