[RM] Fix button press to behave like AlarmReset via bus, and publish it
Previously, one needed to silence every single smoke detector that had
bus alarm individually. With this commit, a device button press on a
smoke detector that has bus alarm causes the alarm to be reset just like
it would have been with an AlarmReset received on the bus, and also the
module sends this event onto the KNX bus by setting AlarmReset to true.
This automatically causes all other smoke detectors with bus alarm to
be silenced.