[RM] Remove 2 minute timeout for attachment to base plate
Previously, we switched on the 12V power supply after 2 minutes, even
if the smoke detector was not mounted on the base plate yet. This does
not make much sense as we have never seen a defective base plate
detection that always yielded "unmounted". We have only seen the
opposite, where a detection yielded "mounted" although it was in fact
not mounted yet.