[RM] Fix periodic sending of `GroupObject::statusAlarm` during delay
When a local alarm starts, we immediately send out `statusAlarm = 1`.
Then, during the delay of a delayed alarm, we did not send out that
communication object anymore, and only started periodic sending after
the delay expired. That is inconsistent.

It was even more obvious when the "also send 0 cyclically" setting was
enabled: Then, we periodically send out `statusAlarm = 0`, stopped the
periodic sending during the delay period, and continued after the delay
completed.

With this change, the `statusAlarm` group object is always sent
periodically when there is a local alarm.