[RM] Merge old files into their current equivalents
For the tests, I actually did those back in December as I had read
`testplan.txt` before I ran the tests. Unfortunately, I forgot to add
two aspects to the test case description back then, so I'm fixing this
up now.

Does not invalidate test results, because I ran the tests the way they
are documented in this commit, and thus the test documentation was
wrong, not the test execution.