[IN8-BCU1] Bugfix for starting with input high
If the application starts with input pins high, the input debouncer was not set up correctly. By calling in loop the debouncer was not set up with valid input level = 1.
the input had to be switched off and on again to fix the correct value.
Now the input debouncer is set up correctly and the value is correct from start.