[ft12] Flip `frameCountBit` on send while still waiting for ft ack
Right now we only have one ft12 out buffer (`ftFrameOutBuffer`).
So we need to send the new frame now, even if we are still waiting for
a ft ack for our last frame.
To send the new frame correct, we need to flip the `frameCountBit` and
fix accordingly the checksum.