[out-cs-bim112] Save time with application load state `LS_LOADED`
During application programming, the ETS checks the current load state of the application.
If this is not `LS_LOADED`, the complete firmware (0x4B20-0x75ED) ~10 kB is transferred.
This firmware is neither used nor stored by our application, so we simply do not need it.
This change saves us ~2 minutes and 17 seconds of programming time during the first ETS application programming.